+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 * Fixed: The path to the Shawl executable was not quoted when it contained spaces.
+* Added: `--priority` option to set the process priority.
 
 ## v1.2.1 (2023-08-10)